Refactor TCP socket client
